+++ test/dummy/test/integration/root_switch_dropdown_test.rb
@@ -35,6 +35,9 @@ class RootSwitchDropdownTest < ActionDispatch::IntegrationTest
     assert_response :success
     assert_includes response.body, workspace.name
     assert_select "body[data-recording-studio-default-layout='true']", count: 1
+    assert_select "header.fp-top-nav", count: 1
+    assert_select "a", text: "Sign out"
+    assert_select "a[href=?]", destroy_user_session_path
   end
 
   test "root switch page renders with the host default layout" do