+++ test/install_generator_test.rb
@@ -233,7 +233,8 @@ class InstallGeneratorTest < Minitest::Test
       '@source "../../../../../../usr/local/bundle/ruby/**/bundler/gems/' \
       'recording_studio_terms_and_conditions-*/app/views/**/*.erb";',
       '@source "../../vendor/bundle/**/flatpack/app/components/**/*.{rb,erb}";',
-      '@source "../../../../../../usr/local/bundle/ruby/**/bundler/gems/flatpack-*/app/components/**/*.{rb,erb}";'
+      '@source "../../../../../../usr/local/bundle/ruby/**/bundler/gems/flatpack-*/app/components/**/*.{rb,erb}";',
+      '@source "../../../../../../usr/local/lib/ruby/gems/**/bundler/gems/flatpack-*/app/components/**/*.{rb,erb}";'
     ]
   end
 end